[gha][PR] Update the PR tests to split tests
This change splits the build and test stages into separate steps.
The purpose of this is to make it easier to make a step as optional.
This change is being added because some distros don't reliably pass the
unit tests in the gate.

A secondary benifit of this is that the jobs can be easily be extended
later by adding additional steps, instead of adding the commands into a
monolithic super step.
Another benifit of this is better reporting, as it is easier to see what
stage the tests fail on, and allow independant test steps to continue to
run even if previous tests failed.